==> code/3_analysis/1_fallow_model_validation.Rmd <==
---
title: "Visualize experiments"
author: "Anna Boser"
date: "2022-11-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(here)
library(scales)
library(data.table)
library(dplyr)
library(stringr)
library(ggplot2)
library(sf)
library(latex2exp)
library(lfe)
library(tidyr)
library(RColorBrewer)

source(here("file_paths.R"))
source(here("helper_functions.R"))
```

```{r}
experiment_name <- "fallow0.05,2_4-18_gb"
experiment_path <- here("data", "4_for_analysis", "ML_outputs",  "experiments", experiment_name)
dataset <- "test" # val or test
tidy <- FALSE
scalar = 12 # mm/month to mm/year

report <- here(experiment_path, "1_report.txt")
if (file.exists(report)) {file.remove(report)} # start fresh every time you render
```

# ET over fallow lands vs predictions
```{r}
# read in the fallow fields predictions
data <- fread(here(experiment_path, paste0("fallow_", dataset, ".csv")))
data[data == -9999] <- NA # this is the na value

if (tidy == FALSE){
  # get a nice column of numeric months
  months <- select(data, names(data)[grepl("month", names(data))])
  data$month <- names(months)[max.col(months)] 
  data$month <- str_extract(data$month, '(?<=_)\\d+') %>% as.numeric()
  data <- select(data, !names(data)[grepl("month_", names(data))])

  # get a nice column of numeric years
  years <- select(data, names(data)[grepl("year", names(data))])
  data$year <- names(years)[max.col(years)] 
  data$year <- str_extract(data$year, '(?<=_)\\d+') %>% as.numeric()
  data <- select(data, !names(data)[grepl("year_", names(data))])
}

write("montly stats" ,file=report,append=TRUE)
MSE <- mean((data$ET_pred - data$ET)^2, na.rm=TRUE)
r2 <- summary(lm(ET~ET_pred, data=data))$r.squared 
Bias <- mean(data$ET_pred - data$ET, na.rm=TRUE)
RMSE <- sqrt(mean((data$ET_pred - data$ET)^2, na.rm=TRUE))
MAE <- mean(abs(data$ET_pred - data$ET), na.rm=TRUE)
line <- paste("R2:", round(r2, 3), "Bias:", round(Bias, 3), "RMSE:", round(RMSE, 3), "MAE:", round(MAE, 3))
write(line,file=report,append=TRUE)
```

### A scatterplot with months averaged out
```{r}
# averaging over different months
data_year <- data %>% group_by(x, y) %>% summarize(ET = mean(ET, na.rm=TRUE)*scalar, 
                                                       ET_pred = mean(ET_pred, na.rm=TRUE)*scalar)

r2 <- summary(lm(ET~ET_pred, data=data_year))$r.squared 
Bias <- mean(data_year$ET_pred - data_year$ET, na.rm=TRUE)
RMSE <- sqrt(mean((data_year$ET_pred - data_year$ET)^2, na.rm=TRUE))
MAE <- mean(abs(data_year$ET_pred - data_year$ET), na.rm=TRUE)
ggplot(data_year) +
  geom_abline(intercept=0,slope=1, color="red") + 
  geom_jitter(aes(x=ET_pred, y=ET), alpha=0.2, size =.1) + 
  theme_bw() + 
  geom_smooth(aes(x=ET_pred, y=ET), method = lm, se = TRUE) +
  # annotate("text", x=3, y=1.5, label= paste("R2:", round(r2, 3), "Bias:", round(Bias, 3))) + 
  coord_fixed() + 
  xlab("Predicted ET (mm/year)") + 
  ylab("Observed ET (mm/year)") +
    xlim(c(0,85*scalar)) + 
    ylim(c(0,85*scalar)) + 
  theme(axis.text.x = element_text(size = 6),
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        legend.title = element_text(size = 7),# Set legend title size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7))

ggsave(here(experiment_path, "Figure_S1.pdf"), 
       width = 12,
       height = 12,
       units = "cm")

write("yearly stats" ,file=report,append=TRUE)
line <- paste("R2:", round(r2, 3), "Bias:", round(Bias, 3), "RMSE:", round(RMSE, 3), "MAE:", round(MAE, 3))
write(line,file=report,append=TRUE)
```

# plot in space
```{r}
study_area <- st_read(study_area_loc) %>% st_transform(st_crs("+proj=longlat +datum=WGS84"))
counties <- st_read(counties_loc) %>% filter(STATEFP == "06") %>% st_transform(st_crs("+proj=longlat +datum=WGS84"))

loc_df <- data %>% group_by(x,y) %>% summarise(Bias = -mean(ag_ET)*scalar, 
                                       ET = mean(ET)*scalar, 
                                       ET_pred = mean(ET_pred)*scalar) %>%
  pivot_longer(cols=c("Bias", "ET", "ET_pred"), names_to = "type", values_to = "ET")

new.labs <- c("Observed ET", "Predicted ET")
names(new.labs) <- c("ET", "ET_pred")

loc_df %>% 
  filter(type %in% c("ET", "ET_pred")) %>%
  ggplot() + 
  geom_sf(data = counties, color=alpha("white",1), fill=alpha("grey",1), size = .2) + 
  geom_sf(data = study_area, fill=alpha("red",0), color = "black", size = .2) + 
  geom_point(aes(x=x, y=y, col=ET), size = .1) + 
  facet_grid(cols = vars(type), labeller = labeller(type = new.labs)) + 
  scale_color_distiller(name="ET (mm/year)", palette="YlGnBu", direction=1) + 
  theme_classic() +
  scale_x_continuous(name = "Longitude", 
                     limits = c(minlong, maxlong),
                     breaks = round(seq(minlong + 1, maxlong + 1, by = 3))) + 
  scale_y_continuous(name = "Latitude", 
                     limits = c(minlat, maxlat),
                     breaks = round(seq(minlat + 1, maxlat + 1, by = 2))) + 
  theme(axis.text.x = element_text(size = 6),
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        legend.title = element_text(size = 7),# Set legend title size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.ticks = element_line())
  
ggsave(here(experiment_path, "Figure_S2.pdf"), 
       width = 18,
       height = 11.2,
       units = "cm")
```
```{r}
lim <- 200 # quantile(abs(loc_df$ET), .40)
print(lim)
loc_df$ET_lim <- ifelse(loc_df$ET>lim, lim, loc_df$ET)
loc_df$ET_lim <- ifelse(loc_df$ET<(-lim), -lim, loc_df$ET)
loc_df %>% filter(type %in% c("Bias")) %>%
  ggplot() + 
  geom_sf(data = counties, color=alpha("white",1), fill=alpha("grey",1), size = .2) + 
  geom_sf(data = study_area, fill=alpha("red",0), color = "black", size = .2) + 
  geom_point(aes(x=x,y=y,col=ET_lim), size = .05) + 
  facet_grid(cols = vars(type)) + 
  scale_color_distiller(name="Bias (mm/year)", palette="RdBu", limits = c(-lim, lim), direction = 1) + 
  xlim(c(minlong, maxlong)) + 
  ylim(c(minlat, maxlat)) + 
  theme(legend.position = c(.87, .84)) + 
  theme_classic() +
  scale_x_continuous(name = "Longitude", 
                     limits = c(minlong, maxlong),
                     breaks = round(seq(minlong + 1, maxlong + 1, by = 3))) + 
  scale_y_continuous(name = "Latitude", 
                     limits = c(minlat, maxlat),
                     breaks = round(seq(minlat + 1, maxlat + 1, by = 2))) + 
  theme(axis.text.x = element_text(size = 6),
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        legend.title = element_text(size = 7),# Set legend title size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.ticks = element_line())

ggsave(here(experiment_path, "Figure_S3.pdf"), 
       width = 11.2,
       height = 11.2,
       units = "cm")
```

# in order to get CIs, I need clustered errors
```{r}
assign_cluster <- function(x, y, dist){
  
  x_size = dist/89 # 1 degree lon (x) = 89km = 89000m
  y_size = dist/111 # 1 degree lat (y) = 111km = 111000m
  
  x_fold = floor(x/x_size)*x_size
  y_fold = floor(y/y_size)*y_size
  
  cv_fold = paste(x_fold, y_fold, sep = ",")
  
  return(cv_fold)
}

data$cluster <- mapply(assign_cluster, data$x, data$y, 75)
```


# Get confidence intervals using clustered standard errors for: 
1. county
2. year
3. month
```{r}
data$year_f <- as.factor(data$year)
data$month_f <- as.factor(data$month)
# get confidence intervals using clustered standard errors
county <- felm(ag_ET~NAME-1 | year_f+month_f | 0 | cluster, data) # -1 means no intercept
year <- felm(ag_ET~year_f-1 | month_f | 0 | cluster, data)
month <- felm(ag_ET~month_f-1 | year_f | 0 | cluster, data)
overall <- felm(ag_ET~1 | 0 | 0 | cluster, data)

summary(overall)
summary(county)
summary(year)
summary(month)
```

```{r}
# extract the mean and ci values from each
ts_from_felm <- function(input_lm){
  ET_df <- as.data.frame(cbind(input_lm$coefficients, input_lm$cse))
  colnames(ET_df) <- c("ET", "se")
  ET_df$min <- ET_df$ET-(1.96*ET_df$se)
  ET_df$max <- ET_df$ET+(1.96*ET_df$se)
  ET_df$month <- as.numeric(substring(rownames(ET_df), 8))
  return(ET_df)
}

ET_df <- ts_from_felm(felm(ET~month_f-1 | 0 | 0 | cluster+year, data))
ET_df$type = "Observed ET"
ET_pred_df <- ts_from_felm(felm(ET_pred~month_f-1 | 0 | 0 | cluster+year, data))
ET_pred_df$type = "Predicted ET"
ts <- rbind(ET_df, ET_pred_df)

# Define a color palette with high contrast
high_contrast_colors <- c("dodgerblue", "firebrick1") # brewer.pal(3, "Set1")

ggplot(ts, aes(x = month)) + 
  geom_line(aes(y = ET, col = type), size = .25) + # make the lines smaller so that they overlap less
  geom_ribbon(aes(ymax = min, ymin = max, fill = type), alpha = .2) + 
  scale_color_manual(values = high_contrast_colors) +  # Use manual colors for lines
  scale_fill_manual(values = high_contrast_colors) +   # Use manual colors for ribbons
  ylab("ET (mm/month)") + 
  xlab("Month") + 
  labs(color=NULL, fill=NULL) + 
  theme_bw() + 
  theme(legend.position = "top", 
    legend.direction = "horizontal", 
    axis.text.x = element_text(size = 6),
    axis.text.y = element_text(size = 6),
    strip.text = element_text(size = 7),
    legend.text = element_text(size = 6), # Set legend text size to 7pt
    legend.title = element_text(size = 7),# Set legend title size to 7pt
    axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
    axis.title.y = element_text(size = 7)) # move legend to the top to allow more space for the plot

ggsave(here(experiment_path, "Figure_S4.pdf"), 
       width = 12,
       height = 8,
       units = "cm")
```
