==> code/3_analysis/2_ag_vs_counter_ET.Rmd <==
---
title: "Analysis"
author: "Anna Boser"
date: '2022-05-02'
output: html_document
---

```{r}
library(here)
library(tidyr)
library(dplyr)
library(ggplot2)
library(rgdal)
library(maptools)
library(data.table)
library(sf)
library(stringr)
library(ggspatial)
library(gstat) # for the variogram
library(egg)

source(here("file_paths.R"))
source(here("helper_functions.R"))
```


# First glance at ET in agriculture 

## Read in the data
```{r}
experiment_name <- "fallow0.05,2_4-18_gb"
experiment_path <- here("data", "4_for_analysis", "ML_outputs",  "experiments", experiment_name)
test <- fread(here(test_data_path, paste0("fallow0.05,2_test.csv"))) # the pixels not used in training or validation
data <- fread(file = here(experiment_path, "agriculture_yearly.csv"))

report <- here(experiment_path, "2_report.txt")
if (file.exists(report)) {file.remove(report)} # start fresh every time you render
```

# clean data
```{r}
data <- data[cropnames != "Urban - residential, commercial, and industrial, unsegregated",] # remove urban
data$cropnames <- ifelse(data$cropnames %in% c("Unclassified fallow", "Idle"), "Fallow", data$cropnames) # rename unclassified
```

## what percent of the variance is due to error? 
```{r}
shrubland_rmse = 19.96 # mm/month
agriculture_rmse = 19.97 # mm/month
model_error_mse = 84.37424 # mm/month

monthly_var_mm_month = (shrubland_rmse^2) + (agriculture_rmse^2) + model_error_mse
yearly_var_mm_month = monthly_var_mm_month/12
yearly_var = yearly_var_mm_month

# a conservative estimate of how much actual variance we've got is simply to remove this from our overall variance
error_attribution = yearly_var/var(data$ag_ET)

# write these in the report
line = paste("Total variance in all the dataset:", var(data$ag_ET))
write(line,file=report,append=TRUE)
line = paste("Total mse from all error sources:", yearly_var)
write(line,file=report,append=TRUE)
line = paste("Proportion of total variance caused by error:", error_attribution)
write(line,file=report,append=TRUE)
```


## variogram
```{r, eval = TRUE}
# plot a variogram 
sample <- data[sample(nrow(data), size = 100000, replace = FALSE),] #subset data to calculate variogram
sample$x <- (sample$x - min(sample$x))*89  # turn x into km 
sample$y <- (sample$y - min(sample$y))*111

vario_ET <- variogram(ET~1, data=sample, locations= ~x+y, cutoff= 100) 
vario_ag_ET <- variogram(ag_ET~1, data=sample, locations= ~x+y, cutoff= 100) 

vario_ET$type = "Total ET"
vario_ag_ET$type = "Agricultural ET"

vario_ET2 <- rbind(vario_ET, vario_ag_ET)

ggplot(vario_ET2) + 
  geom_line(aes(x = dist, y = gamma, color = type)) + 
  geom_point(aes(x = dist, y = gamma, color = type)) + 
  scale_color_manual(values=c("#00618e", "black")) + 
  theme_bw() + 
  theme(legend.position = c(.2, .85), 
        legend.title=element_blank(), 
        axis.text.x = element_text(size = 6),
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7)) + 
  xlab("Distance (km)") + 
  ylab("Gamma")

ggsave(here(experiment_path, "Figure_S12.pdf"), 
       width = 12,
       height = 8,
       units = "cm")
  
```

## histograms 
```{r, eval = FALSE}
# only keep fallow pixels that are in the test set
list <- unique(paste(test$x, test$y, test$year))
nrow(data)
data <- data %>% filter(!(cropnames == "Fallow" & !(paste(x, y, year) %in% list)))
nrow(data)
rm(test)

fallow <- filter(data, cropnames == "Fallow")
hist_data <- filter(data, cropnames != "Fallow")

hist_data <- pivot_longer(hist_data, cols=c(ET, ET_pred, ag_ET), names_to="type", values_to="ET")
hist_data$type <- factor(hist_data$type, levels=c("ET", "ET_pred", "ag_ET"))

# if the the pixel is fallow, make "ET" called "fallow" instead
fallow <- fallow %>% select(x, y, year, ET)
fallow$type <- "fallow"
hist_data <- hist_data %>% select(x, y, year, ET, type)

hist_data <- rbind(hist_data, fallow)

hist_data_full<- hist_data
hist_data <- hist_data_full[sample(1:nrow(hist_data_full), 20000, rep=FALSE),]

ggplot(hist_data) + 
  geom_density(aes(x=ET, fill = type), lwd = 0, alpha = .4) + 
  scale_fill_manual(name="", labels=c("Total ET (on agricultural lands)", "Simulated fallow ET (on agricultural lands)", "Agricultural ET (on agricultural lands)", "Fallow ET (on fallow lands)"), values=c("navyblue", "darkgoldenrod4", "seagreen", "grey")) +
  theme_classic() +  
  theme(legend.position = c(0.75, 0.8),
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank()) +
  xlab("ET (mm/month)")
```


## Make maps of ET, ET_pred, and ag_ET

```{r}
# remove fallow lands
data <- data %>% filter(cropnames != "Fallow")

# Average over years
data <- data %>% group_by(x, y) %>% summarize(ET = mean(ET, na.rm=TRUE), 
                                                ET_pred = mean(ET_pred, na.rm=TRUE), 
                                                ag_ET = ET-ET_pred)
```

```{r}
data <- pivot_longer(data, cols=c(ET, ET_pred, ag_ET), names_to="type", values_to="ET")
data$type <- factor(data$type, levels=c("ET", "ET_pred", "ag_ET"))
```

```{r}
# study area and counties for plotting
study_area <- st_read(study_area_loc) %>% st_transform(st_crs("+proj=longlat +datum=WGS84"))
counties <- st_read(counties_loc) %>% filter(STATEFP == "06") %>% st_transform(st_crs("+proj=longlat +datum=WGS84"))
```

```{r}
scalar = 12 # mm/month to mm/year
# full_data <- data
# data <- full_data[sample(nrow(full_data), size = 100000, replace = FALSE),] # subset for fast testing

new.labs <- c("Total ET", "Naturally-occurring ET", "Agricultural ET")
names(new.labs) <- c("ET", "ET_pred", "ag_ET")

ggplot() + 
  geom_sf(data = counties, fill = "grey", color=alpha("white",1), size = .2) + 
  geom_raster(data = data, aes(x=x, y=y, fill=ET*scalar)) +
  facet_grid(cols=vars(type), labeller = labeller(type = new.labs)) +
  scale_fill_gradientn(name="ET (mm/year)", colours = c("white", "#FFFFCC", "#A1DAB4",  "#41B6C4", "#2C7FB8", "#253494", "#081D58", "#081D58"), limits = c(-391, 2346), breaks = seq(0, 2000, by = 1000)) +
  geom_sf(data = study_area, fill=alpha("red",0), color = "black", size = .2) + 
  xlim(c(minlong, maxlong)) + 
  ylim(c(minlat, maxlat)) + 
  annotation_scale(data = tibble::tibble(type = as.factor("ET")), text_cex=.6, height=unit(0.2, "cm")) + 
  theme_classic() +
  scale_x_continuous(name = "Longitude", 
                     limits = c(minlong, maxlong),
                     breaks = round(seq(minlong + 1, maxlong + 1, by = 3))) + 
  scale_y_continuous(name = "Latitude", 
                     limits = c(minlat, maxlat),
                     breaks = round(seq(minlat + 1, maxlat + 1, by = 2))) + 
  theme(legend.position = c(0.934, 0.867), 
        legend.direction = "horizontal",
        axis.text.x = element_text(size = 6),
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        legend.title = element_text(size = 7),# Set legend title size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.ticks = element_line(), 
        legend.key.width = unit(dev.size()[1] / 40, "inches")) + 
  guides(fill = guide_colourbar(title.position="top", title.hjust = 0.5, override.aes = list(size = 0.5))) 

ggsave(here(experiment_path, "Figure_1.pdf"), 
       width = 18,
       height = 9,
       units = "cm")
```

```{r}
# extra stuff to photoshop later
ggplot() + 
  geom_sf(data = counties, fill = "grey", color=alpha("white",0), size = .2) + 
  geom_sf(data = study_area, fill=alpha("red",0), color = "black", size = .5) + 
  theme_void() +
  geom_rect(aes(xmin=minlong-.2, xmax=maxlong+.2, ymin=minlat-.2, ymax=maxlat+.2), color="black", fill="transparent", size = .5)

ggsave(here(experiment_path, "Figure_1_insert.pdf"), 
       width = 2.75,
       height = 2.75,
       units = "cm")
```
