==> code/3_analysis/3_crops_comparison.Rmd <==
---
title: "crops comparison"
author: "Anna Boser"
date: '2022-05-05'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(sp)
library(here)
library(stringr)
library(sf)
#library(tmap)
library(data.table)
library(tidyr)
library(dplyr)
library(ggplot2)
library(latex2exp)
library(lfe)

source(here("file_paths.R"))
source(here("helper_functions.R"))
```

## Read in the data
```{r}
experiment_name <- "fallow0.05,2_4-18_gb"
experiment_path <- here("data", "4_for_analysis", "ML_outputs",  "experiments", experiment_name)
data <- fread(file = here(experiment_path, "agriculture_yearly.csv"))
# test <- fread(here(test_data_path, paste0("fallow0,2_test.csv"))) # the pixels not used in training or validation 
test <- fread(here(test_data_path, paste0("fallow0.05,2_test.csv"))) # the pixels not used in training or validation 

report <- here(experiment_path, "3_report.txt")
if (file.exists(report)) {file.remove(report)} # start fresh every time you render
```

# clean data
```{r}
data <- filter(data, coverage_fraction>.5)

data <- data[cropnames != "",] # only keep data that has crop info (get rid of years without dwr data)
data <- data[cropnames != "Urban - residential, commercial, and industrial, unsegregated",] # remove urban
data$cropnames <- ifelse(data$cropnames %in% c("Unclassified fallow", "Idle"), "Fallow", data$cropnames) # rename unclassified fallow to just fallow

# only keep fallow pixels that are in the test set
list <- unique(paste(test$x, test$y, test$year))
nrow(data)
data <- data %>% filter(!(cropnames == "Fallow" & !(paste(x, y, year) %in% list)))
nrow(data)
rm(test)

data$subcropnames <- ifelse(data$subcropnames == "", data$cropnames, data$subcropnames)

data <- filter(data,  subcropnames != "Greenhouse")

# combine potatoes and "potatoes or sweet potato"
data$subcropnames <- ifelse(data$subcropnames %in% c("Potatoes", "Potato or Sweet potato"), "Potato or Sweet potato", data$subcropnames)
```

```{r}
# create clusters based on location and year. Function creates clusters of size dist km. 
data$cluster <- mapply(assign_cluster, data$x, data$y, 75)

# scale to mm/year instead of mm/month
scalar = 12 # turns mm/month to mm/year
data <- scale(data, scalar)

# control for variation across years
data <- year_control(data)
```


## Numbers
```{r}
#percent of total ET that is ag water consumption
total_ET <- mean(filter(data, cropnames != "Fallow")$ET)
ag_ET_mean <- felm(ag_ET~1 | 0 | 0 | cluster, filter(data, cropnames != "Fallow"))
summary(ag_ET_mean)

line <- print_CI("avg ag ET:", ag_ET_mean$coefficients[1], ag_ET_mean$cse)
write(line,file=report,append=TRUE)

line <- print_CI("percent ag ET:", (ag_ET_mean$coefficients[1]/total_ET)*100, (ag_ET_mean$cse/total_ET)*100)
write(line,file=report,append=TRUE)

#percent of total ET that is natural water consumption
nat_ET_mean <- felm(ET_pred~1 | 0 | 0 | cluster, filter(data, cropnames != "Fallow"))
summary(nat_ET_mean)

line <- print_CI("avg nat ET:", nat_ET_mean$coefficients[1], nat_ET_mean$cse)
write(line,file=report,append=TRUE)

line <- print_CI("percent ag ET:", (nat_ET_mean$coefficients[1]/total_ET)*100, (nat_ET_mean$cse/total_ET)*100)
write(line,file=report,append=TRUE)

# check if the variation of naturally-occurring ET across space is significant
summary(felm(ET_pred~x*y | 0  | 0 | cluster, filter(data, cropnames == "Fallow")))

# variation within crop types
sds <- data %>% group_by(cropnames) %>% summarize(sd = sd(ag_ET), 
                                                  IQR = IQR(ag_ET))
print(sds, n = nrow(sds))
line <- paste("active sds average (sd within crops):", mean(filter(sds, cropnames != "Fallow")$sd))
write(line,file=report,append=TRUE)
```

##################
# compare crops
```{r}
# get confidence intervals using clustered standard errors
ET_c <- felm(ET~cropnames-1 | 0  | 0 | cluster, data) # -1 means no intercept
# ET_c %>% summary()
ET_df <- df_from_felm(ET_c)
rm(ET_c)

ag_ET_c <- felm(ag_ET~cropnames-1 | 0  | 0 | cluster, data)
# ag_ET_c %>% summary()
ag_ET_df <- df_from_felm(ag_ET_c)
rm(ag_ET_c)

ET_pred_c <- felm(ET_pred~cropnames-1 | 0  | 0 | cluster, data)
# ET_pred_c %>% summary()
ET_pred_df <- df_from_felm(ET_pred_c)
rm(ET_pred_c)

ET_df$type = "ET"
ag_ET_df$type = "ag_ET"
ET_pred_df$type = "ET_pred"

bardata <- rbind(ET_df, ET_pred_df, ag_ET_df)
bardata$type <- as.factor(bardata$type)
bardata$type <- relevel(bardata$type, 'ET_pred')

# sd across crop types
print("sd across crop types:")
sd(filter(bardata, type == "ag_ET")$ET)
```

### Plot

California’s irrigated pasturelands—including valley, foothill, and mountain meadow pastures— account for nearly 500,000 acres across the state, and ranks third among agricultural water users. https://rangelands.ucdavis.edu/ipep/

```{r}
print(bardata %>% filter(type == "ag_ET"))
line <- print_CI("Deciduous fruits and nuts ag ET", mean = filter(bardata, type == "ag_ET")$ET[2], se = filter(bardata, type == "ag_ET")$se[2])
write(line,file=report,append=TRUE)
line <- print_CI("Grain and hay crops ag ET", mean = filter(bardata, type == "ag_ET")$ET[5], se = filter(bardata, type == "ag_ET")$se[5])
write(line,file=report,append=TRUE)
```
```{r}
boxplot <- filter(data, cropnames != "Young Perennial") %>%
  group_by(cropnames) %>%
  summarise(ET = 1,
            percentile_05 = quantile(ag_ET, .05), 
            percentile_25 = quantile(ag_ET, .25),
            percentile_50 = quantile(ag_ET, .5), 
            percentile_75 = quantile(ag_ET, .75), 
            percentile_95 = quantile(ag_ET, .95))
```


```{r}

filter(bardata, type == "ag_ET", cropnames != "Young Perennial") %>% 
  ggplot(aes(x = reorder(cropnames, ET))) + 
  geom_linerange(data=filter(bardata, type == "ag_ET", cropnames != "Young Perennial"), aes(ymax = min, ymin = max), color = "#00618e") + 
  geom_boxplot(data=boxplot, aes(lower = percentile_25, upper = percentile_75, middle = percentile_50, ymin = percentile_05, ymax = percentile_95), width=.7, color = "grey50", stat="identity") + 
  geom_col(data = filter(bardata, type != "ET", cropnames != "Young Perennial"), aes(y = ET, fill = type), alpha = .6) +
  scale_fill_manual(values=c(ag_ET="#00618e", ET_pred="#ede1c5"), breaks=c("ag_ET","ET_pred"), labels = c("Agricultural ET", "Naturally-occurring ET")) + 
  geom_col(data = filter(bardata, type == "ET", cropnames != "Young Perennial"), aes(y = ET, color = type), size=.3, alpha = 0) +
  scale_color_manual(values=c("black"), labels = c("Total ET")) + 
  # geom_linerange(data = filter(bardata, type == "ET", cropnames != "Young Perennial"), aes(ymax = min, ymin = max), size=1.1, color = "navyblue") + 
  geom_linerange(aes(ymax = min, ymin = max), size=1.1, color = "#00618e") + 
  ylab("mm/year") + 
  theme_bw() + 
  labs(fill='') + 
  # ylim(c(-35, 165)) + 
  theme(axis.text.x = element_text(angle = 30, hjust=1, size=7), 
        axis.title.x=element_blank(), 
        axis.ticks.x=element_blank(), 
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 7), # Set legend text size to 7pt
        axis.title.y = element_text(size = 7),
        legend.position = "top", 
        legend.direction="horizontal", 
        legend.title=element_blank())

ggsave(here(experiment_path, "Figure_2.pdf"), 
       width = 18,
       height = 11.2,
       units = "cm")
```

```{r}
# Make a dataset to save and publish
crop_group <- bardata %>% filter(type == "ag_ET", cropnames != "Young Perennial")
crop_group$type <- NULL
crop_group$se <- NULL
crop_group$agricultural_ET <- crop_group$ET
crop_group$ET <- NULL
crop_group$agricultural_ET_min <- crop_group$min
crop_group$min <- NULL
crop_group$agricultural_ET_max <- crop_group$max
crop_group$max <- NULL

boxplot$ET <- NULL

crop_group <- merge(crop_group, boxplot, by = "cropnames")
crop_group$crop_group <- crop_group$cropnames
crop_group$cropnames <- NULL

fwrite(crop_group, here(experiment_path, "crop_group_table.csv"))
```


# scrap plots
```{r, eval = FALSE, render=FALSE}
# The original bardata without boxplot

ggplot(filter(bardata, type == "ag_ET"), aes(x = reorder(cropnames, ET))) +
  geom_linerange(data = filter(bardata, type == "ag_ET"), aes(ymax = min, ymin = max), color = "#00618e") +
  geom_col(data = filter(bardata, type != "ET"), aes(y = ET, fill = type), alpha = .6) +
  scale_fill_manual(values=c(ag_ET="#00618e", ET_pred="#ede1c5"), breaks=c("ag_ET","ET_pred"), labels = c("Agricultural ET", "Simulated natural ET")) +
  geom_col(data = filter(bardata, type == "ET"), aes(y = ET, color = type), alpha = 0) +
  scale_color_manual(values=c("navyblue"), labels = c("Total ET")) +
  geom_linerange(data = filter(bardata, type == "ET"), aes(ymax = min, ymin = max), color = "navyblue") +
  geom_linerange(data = filter(bardata, type == "ag_ET"), aes(ymax = min, ymin = max), color = "#00618e") +
  ylab("mm/year") +
  theme_bw() +
  labs(fill='') +
  theme(axis.text.x = element_text(angle = 30, hjust=1),
        axis.title.x=element_blank(),
        axis.ticks.x=element_blank(),
        legend.position = "top",
        legend.direction="horizontal",
        legend.title=element_blank())

# boxplot to show spread in how much a given crop consumes

ggplot(filter(bardata, type == "ag_ET"), aes(x = reorder(cropnames, ET))) + 
  geom_linerange(aes(ymax = min, ymin = max), color = "#00618e") +
  geom_boxplot(data=data, aes(y=ag_ET), outlier.size = .5, outlier.alpha = .002) + 
  geom_col(aes(y = ET), fill = "#00618e", alpha = 0.6) + 
  geom_linerange(aes(ymax = min, ymin = max), color = "#00618e") +
  theme_bw() + 
  labs(fill='') + 
  ylab("Agricultural ET (mm/year)") + 
  theme(axis.text.x = element_text(angle = 30, hjust=1), 
        axis.title.x=element_blank(), 
        axis.ticks.x=element_blank(), 
        legend.position = "top", 
        legend.direction="horizontal", 
        legend.title=element_blank()) 
```

```{r}
# are ET_pred_c different from each other? 
felm(ET_pred~cropnames | 0  | 0 | cluster, data) %>% summary
# how much of the variation is explained by crop type? 
felm(ag_ET~cropnames | 0  | 0 | cluster, data) %>% summary
# what about when you control for climate
felm(ag_ET~cropnames+PET+Elevation+Aspect+Slope+TWI+Soil | 0  | 0 | cluster, data) %>% summary
```

Fallow ET_pred is significantly lower and Rice is significantly higher. 

# compare subcrops
```{r}
# get rid of classes I'm not that interested in 
data <- filter(data,  !(subcropnames %in% c("Miscellaneous deciduous", "Miscellaneous field", "Miscellaneous grain and hay", "Greenhouse", "Miscellaneous grasses", "Miscellaneous subtropical fruit", "Miscellaneous truck")))

# variation within crop types
sds <- data %>% group_by(subcropnames) %>% summarize(sd = sd(ag_ET), 
                                                  IQR = IQR(ag_ET))
print(sds, n = nrow(sds))
line <- paste("active sds average (sd within subcrops):", mean(filter(sds, subcropnames != "Fallow")$sd))
write(line,file=report,append=TRUE)
```


```{r}
# get confidence intervals using clustered standard errors
ET_c <- felm(ET~subcropnames-1 | 0  | 0 | cluster, data) # -1 means no intercept
# ET_c %>% summary()
ET_df <- df_from_felm(ET_c, 13, "subcropnames")
rm(ET_c)

ag_ET_c <- felm(ag_ET~subcropnames-1 | 0  | 0 | cluster, data)
# ag_ET_c %>% summary()
ag_ET_df <- df_from_felm(ag_ET_c, 13, "subcropnames")
rm(ag_ET_c)

ET_pred_c <- felm(ET_pred~subcropnames-1 | 0  | 0 | cluster, data)
# ET_pred_c %>% summary()
ET_pred_df <- df_from_felm(ET_pred_c, 13, "subcropnames")
rm(ET_pred_c)

ET_df$type = "ET"
ag_ET_df$type = "ag_ET"
ET_pred_df$type = "ET_pred"

bardata <- rbind(ET_df, ET_pred_df, ag_ET_df)
bardata$type <- as.factor(bardata$type)
bardata$type <- relevel(bardata$type, 'ET_pred')

# sd across crop types
line <- paste("sd across crop types:", sd(filter(bardata, type == "ag_ET")$ET))
write(line,file=report,append=TRUE)
```

```{r}
print(bardata %>% filter(type == "ag_ET"))
line <- print_CI("Almonds", mean = filter(bardata, type == "ag_ET")$ET[2], se = filter(bardata, type == "ag_ET")$se[2])
write(line,file=report,append=TRUE)
```

### Plot

California’s irrigated pasturelands—including valley, foothill, and mountain meadow pastures— account for nearly 500,000 acres across the state, and ranks third among agricultural water users. https://rangelands.ucdavis.edu/ipep/

```{r}
boxplot <- filter(data, subcropnames != "Young Perennial") %>%
  group_by(subcropnames) %>%
  summarise(ET = 1,
            percentile_05 = quantile(ag_ET, .05), 
            percentile_25 = quantile(ag_ET, .25),
            percentile_50 = quantile(ag_ET, .5), 
            percentile_75 = quantile(ag_ET, .75), 
            percentile_95 = quantile(ag_ET, .95))

ggplot(filter(bardata, type == "ag_ET", subcropnames != "Young Perennial"), aes(x = reorder(subcropnames, ET))) + 
  geom_linerange(data = filter(bardata, type == "ag_ET", subcropnames != "Young Perennial"), aes(ymax = min, ymin = max), color = "#00618e") + 
  geom_boxplot(data=boxplot, aes(lower = percentile_25, upper = percentile_75, middle = percentile_50, ymin = percentile_05, ymax = percentile_95), color = "grey50", stat="identity") + 
  geom_col(data = filter(bardata, type != "ET", subcropnames != "Young Perennial"), aes(y = ET, fill = type), alpha = .6) +
  scale_fill_manual(values=c(ag_ET="#00618e", ET_pred="#ede1c5"), breaks=c("ag_ET","ET_pred"), labels = c("Agricultural ET", "Naturally-occurring ET")) + 
  geom_col(data = filter(bardata, type == "ET", subcropnames != "Young Perennial"), aes(y = ET, color = type), alpha = 0, size = 0.3) +
  scale_color_manual(values=c("black"), labels = c("Total ET")) + 
  # geom_linerange(data = filter(bardata, type == "ET", subcropnames != "Young Perennial"), aes(ymax = min, ymin = max), size = 1, color = "navyblue") + 
  geom_linerange(data = filter(bardata, type == "ag_ET", subcropnames != "Young Perennial"), aes(ymax = min, ymin = max), size = 1, color = "#00618e") + 
  ylab("mm/year") + 
  theme_bw() + 
  labs(fill='') + 
  # ylim(c(-45, 175)) + 
  theme(axis.text.x = element_text(angle = 90, hjust=1, vjust=.5, size = 7), 
        axis.title.x=element_blank(), 
        axis.text.y = element_text(size = 7),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 7), # Set legend text size to 7pt
        axis.title.y = element_text(size = 7),
        legend.position = "top", 
        legend.direction="horizontal", 
        legend.title=element_blank())

ggsave(here(experiment_path, "Figure_S5.pdf"), 
       width = 18,
       height = 11.2,
       units = "cm")
```

```{r}
# Make a dataset to save and publish
crop_type <- bardata %>% filter(type == "ag_ET", subcropnames != "Young Perennial")
crop_type$type <- NULL
crop_type$se <- NULL
crop_type$agricultural_ET <- crop_type$ET
crop_type$ET <- NULL
crop_type$agricultural_ET_min <- crop_type$min
crop_type$min <- NULL
crop_type$agricultural_ET_max <- crop_type$max
crop_type$max <- NULL

boxplot$ET <- NULL

crop_type <- merge(crop_type, boxplot, by = "subcropnames")
crop_type$crop_type <- crop_type$subcropnames
crop_type$cropnames <- NULL

fwrite(crop_type, here(experiment_path, "crop_type_table.csv"))
```


# scrap plots
```{r, eval = FALSE, render=FALSE}
ggplot(filter(bardata, type == "ag_ET"), aes(x = reorder(subcropnames, ET))) + 
  geom_linerange(data = filter(bardata, type == "ag_ET"), aes(ymax = min, ymin = max), color = "#00618e") + 
  geom_col(data = filter(bardata, type != "ET"), aes(y = ET, fill = type), alpha = .6) +
  scale_fill_manual(values=c(ag_ET="#00618e", ET_pred="#ede1c5"), breaks=c("ag_ET","ET_pred"), labels = c("Agricultural ET", "Simulated natural ET")) + 
  geom_col(data = filter(bardata, type == "ET"), aes(y = ET, color = type), alpha = 0) +
  scale_color_manual(values=c("navyblue"), labels = c("Total ET")) + 
  geom_linerange(data = filter(bardata, type == "ET"), aes(ymax = min, ymin = max), color = "navyblue") + 
  geom_linerange(data = filter(bardata, type == "ag_ET"), aes(ymax = min, ymin = max), color = "#00618e") + 
  ylab("mm/year") + 
  theme_bw() + 
  labs(fill='') + 
  theme(axis.text.x = element_text(angle = 90, hjust=1, vjust=.5), 
        axis.title.x=element_blank(), 
        # axis.ticks.x=element_blank(), 
        legend.position = "top", 
        legend.direction="horizontal", 
        legend.title=element_blank())


# boxplot to show spread in how much a given crop consumes

ggplot(filter(bardata, type == "ag_ET"), aes(x = reorder(subcropnames, ET))) + 
  geom_linerange(aes(ymax = min, ymin = max), color = "#00618e") +
  geom_boxplot(data=data, aes(y=ag_ET), outlier.size = .5, outlier.alpha = .002) + 
  geom_col(aes(y = ET), fill = "#00618e", alpha = 0.6) + 
  geom_linerange(aes(ymax = min, ymin = max), color = "#00618e") +
  theme_bw() + 
  labs(fill='') + 
  ylab("Agricultural ET (mm/year)") + 
  theme(axis.text.x = element_text(angle = 90, hjust=1, vjust=.5), 
        axis.title.x=element_blank(), 
        # axis.ticks.x=element_blank(), 
        legend.position = "top", 
        legend.direction="horizontal", 
        legend.title=element_blank()) 

# Total ET 
ggplot(filter(bardata, type == "ET"), aes(x = reorder(subcropnames, ET))) + 
  geom_linerange(aes(ymax = min, ymin = max), color = "navyblue") +
  geom_boxplot(data=data, aes(y=ET), outlier.size = .5, outlier.alpha = .002) + 
  geom_col(aes(y = ET), fill = "navyblue", alpha = 0.6) + 
  geom_linerange(aes(ymax = min, ymin = max), color = "navyblue") +
  theme_bw() + 
  labs(fill='') + 
  ylab("Total ET (mm/year)") + 
  theme(axis.text.x = element_text(angle = 90, hjust=1, vjust=.5), 
        axis.title.x=element_blank(), 
        # axis.ticks.x=element_blank(), 
        legend.position = "top", 
        legend.direction="horizontal", 
        legend.title=element_blank()) 
```


```{r}
# are ET_pred_c different from each other? 
felm(ET_pred~subcropnames | 0  | 0 | cluster, data) %>% summary
# how much of the variation is explained by crop type? 
felm(ag_ET~subcropnames | 0  | 0 | cluster, data) %>% summary
# what about when you control for climate
felm(ag_ET~subcropnames+PET+Elevation+Aspect+Slope+TWI+Soil | 0  | 0 | cluster, data) %>% summary

# what percent of within-crop variation do we account for with climate? 
mean_by_crop <- data %>% group_by(subcropnames) %>% summarize(crop_mean = mean(ag_ET))
data <- merge(data, mean_by_crop, by = c("subcropnames"))
data$within_crop_var <- data$ag_ET - data$crop_mean

mod <- felm(within_crop_var~PET+Elevation+Aspect+Slope+TWI+Soil | 0  | 0 | cluster, data) %>% summary
line <- paste("percent within crop variation accounted for by PET+Elevation+Aspect+Slope+TWI+Soil:", mod$r2)
write(line,file=report,append=TRUE)
```
