==> code/3_analysis/4_scenarios.Rmd <==
---
title: "5_gsa_analysis"
author: "Anna Boser"
date: "2023-03-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(sp)
library(here)
library(stringr)
library(sf)
library(data.table)
library(tidyr)
library(dplyr)
library(ggplot2)
library(latex2exp)
library(ggspatial)
library(lfe)

source(here("file_paths.R"))
source(here("helper_functions.R"))
```

## Read in the data
```{r}
experiment_name <- "fallow0.05,2_4-18_gb"
# experiment_name <- "fallow0,2_4-23_gb"
experiment_path <- here("data", "4_for_analysis", "ML_outputs",  "experiments", experiment_name)
data <- fread(file = here(experiment_path, "agriculture_yearly.csv"))

# study area and counties for plotting
study_area <- st_read(study_area_loc) %>% st_transform(st_crs("+proj=longlat +datum=WGS84"))
counties <- st_read(counties_loc) %>% filter(STATEFP == "06") %>% st_transform(st_crs("+proj=longlat +datum=WGS84"))

report <- here(experiment_path, "4_report.txt")
if (file.exists(report)) {file.remove(report)} # start fresh every time you render
```

# only keep 2019 since this is the only year where I have enough information about orchard age
```{r}
data <- filter(data, year == 2019)
```

# clean data
```{r}
data <- filter(data, coverage_fraction>.5)

data <- data[cropnames != "",] # only keep data that has crop info (get rid of years without dwr data)
data <- data[cropnames != "Urban - residential, commercial, and industrial, unsegregated",] # remove urban
data$cropnames <- ifelse(data$cropnames %in% c("Unclassified fallow", "Idle"), "Fallow", data$cropnames) # rename unclassified fallow to just fallow

# remove all fallow pixels
data <- filter(data, cropnames != "Fallow")

data$subcropnames <- ifelse(data$subcropnames == "", data$cropnames, data$subcropnames)

data$year <- as.factor(data$year)

data <- filter(data,  subcropnames != "Greenhouse")

# combine potatoes and "potatoes or sweet potato"
data$subcropnames <- ifelse(data$subcropnames %in% c("Potatoes", "Potato or Sweet potato"), "Potato or Sweet potato", data$subcropnames)

# # control for variation across years (don't need to do this since I'm only looking at 2019 anyways)
# data <- year_control(data)

```

```{r}
# add basin info
data <- add_basin(data)

data <- filter(data, !is.na(subbasin_n)) # remove pixels that are not in a subbasin
data <- filter(data, subbasin_n != "CAPAY VALLEY") # this is tiny and not really in the CV
data <- filter(data, cropnames != "Young Perennial") # remove young perennials because you can't actually switch to these
```


## Only keep orchards that are not young or old
```{r}
# read in all crop information
get_crop <- function(year){
  crop <- fread(here(crops_dwr_table_path, paste0("crops_all_dwr_fallow", year, ".csv")))
  crop$year <- year
  return(crop)
}
crop_data <- rbindlist(lapply(c(2014, 2016, 2018, 2020), get_crop), fill = TRUE)

# get a list of locations where a location was "Young Perennial" in any of the other years
YP <- filter(crop_data, cropnames == "Young Perennial")
old_young_locations <- unique(paste(sprintf("%.7f", YP$x), sprintf("%.7f", YP$y)))

# remove instances from the data where a deciduous fruits & nuts or citrus & subtropical pixel was YP in any of the other years
data$loc <- paste(sprintf("%.7f", data$x), sprintf("%.7f", data$y))
data <- filter(data, !(cropnames %in% c("Deciduous fruits and nuts", "Citrus and subtropical") & loc %in% old_young_locations))
```
## control for climate and other physical aspects
```{r}
# there are some NA values and this keeps the linear models from making predictions, so we toss these
data <- filter(data, !is.na(PET), !is.na(Elevation), !is.na(Aspect), !is.na(Slope), !is.na(TWI), !is.na(Soil)) 

# remove climate variation
data$ag_ET <- data$ag_ET - (lm(ag_ET~PET+Elevation+Aspect+Slope+TWI+Soil, data)$fitted.values - mean(data$ag_ET))
```

# get medians, means, etc. 

```{r}
# stuff by basin
# get the mean consumption for each crop by basin and year
crop_means_basin <- data %>% group_by(subcropnames, subbasin_n, year) %>% summarize(crop_mean_basin = mean(ag_ET))
crop_means_basin$crop_mean_basin <- ifelse(crop_means_basin$crop_mean_basin <0, 0, crop_means_basin$crop_mean_basin)

# merge to the larger dataset
data <- merge(data, crop_means_basin, by = c("subcropnames", "subbasin_n", "year"))

# get the median consumption for each crop by basin
crop_medians_basin <- data %>% group_by(subcropnames, subbasin_n, year) %>% summarize(crop_median_basin = median(ag_ET))
crop_medians_basin$crop_median_basin <- ifelse(crop_medians_basin$crop_median_basin <0, 0, crop_medians_basin$crop_median_basin)

# merge to the larger dataset
data <- merge(data, crop_medians_basin, by = c("subcropnames", "subbasin_n", "year")) 

# pick the median crop for each basin and year. This is the median crop grown arranged by mean ag ET
basin_median_crop <- data %>% group_by(subbasin_n, year) %>% summarize(basin_median_crop = median(crop_mean_basin))
data <- merge(data, basin_median_crop, by = c("subbasin_n", "year"))

# pick the minimum crop for each basin. 
basin_min_crop <- data %>% group_by(subbasin_n, year) %>% summarize(basin_min_crop = min(crop_mean_basin))
data <- merge(data, basin_min_crop, by = c("subbasin_n", "year"))

# 95th percentile consumers in each basin
basin_95 <- data %>% group_by(subbasin_n, year) %>% summarize(basin_95 = quantile(ag_ET, probs = .95))
data <- merge(data, basin_95, by = c("subbasin_n", "year"))
```

# scenarios 
```{r}
# Crop-switching to the median consuming crop
data$med_crop_switch <- ifelse(data$crop_mean_basin > data$basin_median_crop, data$basin_median_crop, data$crop_mean_basin)

# Crop-switching to the minimum consuming crop
data$min_crop_switch <- ifelse(data$crop_mean_basin > data$basin_min_crop, data$basin_min_crop, data$crop_mean_basin)

# same crop -- median consumption
data$median_same_crop <- ifelse(data$ag_ET > data$crop_median_basin, data$crop_median_basin, data$ag_ET)

# fallow 5% of lands
data$fallow <- ifelse(data$ag_ET > data$basin_95, 0, data$ag_ET)
```

```{r}
# get the percent saved for each intervention
scenarios <- data %>% 
  summarise(med_crop_switch = 100 - mean(med_crop_switch)/mean(ag_ET)*100, 
            min_crop_switch = 100 - mean(min_crop_switch)/mean(ag_ET)*100, 
            median_same_crop = 100 - mean(median_same_crop)/mean(ag_ET)*100, 
            fallow = 100 - mean(fallow)/mean(ag_ET)*100) %>%
  pivot_longer(cols = c("med_crop_switch", "min_crop_switch", "median_same_crop", "fallow"), 
               names_to = "Intervention", 
               values_to = "Percent_savings")

line <- print(scenarios, n = nrow(scenarios))
fwrite(line,file=report,append=TRUE)

# scenarios by basin for plotting
scenarios <- data %>% 
  group_by(subbasin_n) %>% 
  summarise(med_crop_switch = 100 - mean(med_crop_switch)/mean(ag_ET)*100, 
            min_crop_switch = 100 - mean(min_crop_switch)/mean(ag_ET)*100, 
            median_same_crop = 100 - mean(median_same_crop)/mean(ag_ET)*100, 
            fallow = 100 - mean(fallow)/mean(ag_ET)*100) %>%
  pivot_longer(cols = c("med_crop_switch", "min_crop_switch", "median_same_crop", "fallow"), 
               names_to = "Intervention", 
               values_to = "Percent_savings")

# basins 
basin_poly <- st_read(basin_loc) 
# filter to only keep CV basins
basin_poly <- basin_poly %>% filter(basin_name %in% c("SACRAMENTO VALLEY", "SAN JOAQUIN VALLEY"))
basin_poly <- filter(basin_poly, subbasin_n != "CAPAY VALLEY") # this is tiny and not really in the CV
scenarios <- merge(basin_poly, scenarios, by = "subbasin_n", all.x = TRUE)
```

```{r}

new.labs <- c("Crop-switching", "Switch to minimum crop", "Farming practices", "Fallowing 5%")
names(new.labs) <- c("med_crop_switch", "min_crop_switch", "median_same_crop", "fallow")

scenarios$Intervention <- factor(scenarios$Intervention, levels = c("med_crop_switch", "min_crop_switch", "median_same_crop", "fallow"))

scenarios$Percent_savings_color <- pmin(scenarios$Percent_savings, 20)

filter(scenarios, !(Intervention %in% c("min_crop_switch"))) %>% 
#scenarios %>%
  ggplot() + 
  geom_sf(data = counties, fill = "grey", color=alpha("white",1), size = .2) + 
  geom_sf(aes(fill = Percent_savings_color), color=alpha("black",.3), size = .2) + 
  facet_grid(cols = vars(Intervention), labeller = labeller(Intervention = new.labs)) +
  scale_fill_distiller(palette="Blues", direction=1, name = "% Savings", labels = c("0", "5", "10", "15", ">20"), breaks = seq(0, 20, by = 5)) + 
  geom_sf(data = study_area, fill=alpha("red",0), color = "black", size = .2) + 
  xlim(c(minlong, maxlong)) + 
  ylim(c(minlat, maxlat)) + 
  theme_classic() +
  annotation_scale(data = tibble::tibble(Intervention = as.factor("med_crop_switch")), text_cex=.6, height=unit(0.2, "cm")) +
  scale_x_continuous(name = "Longitude", 
                     limits = c(minlong, maxlong),
                     breaks = round(seq(minlong + 1, maxlong + 1, by = 3))) + 
  scale_y_continuous(name = "Latitude", 
                     limits = c(minlat, maxlat),
                     breaks = round(seq(minlat + 1, maxlat + 1, by = 2))) + 
  theme(legend.position = c(0.934, 0.867), 
        legend.direction = "horizontal",
        axis.text.x = element_text(size = 6),
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        legend.title = element_text(size = 7),# Set legend title size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.ticks = element_line(), 
        legend.key.width = unit(dev.size()[1] / 40, "inches")) + 
  guides(fill = guide_colourbar(title.position="top", title.hjust = 0.5, override.aes = list(size = 0.5))) 

ggsave(here(experiment_path, "Figure_3.pdf"), 
       width = 18,
       height = 8.5,
       units = "cm")
```
