==> code/3_analysis/5_irrigation_comparison.Rmd <==
---
title: "irrigation comparison"
author: "Anna Boser"
date: '2022-05-04'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(here)
library(ggpubr)
library(stringr)
library(sf)
library(data.table)
library(tidyr)
library(dplyr)
library(latex2exp)
library(ggplot2)
library(lfe)
library(rgdal)
library(ggspatial)
library(ggnewscale) # allows you to plot multiple things with different scales on one ggplot

source(here("file_paths.R"))
source(here("helper_functions.R"))
```


## Read in the data
```{r}
experiment_name <- "fallow0.05,2_4-18_gb"
# experiment_name <- "fallow0,2_4-23_gb"
experiment_path <- here("data", "4_for_analysis", "ML_outputs",  "experiments", experiment_name)
ET_year <- NA
irr_year <- NA

report <- here(experiment_path, "5_report.txt")
if (file.exists(report)) {file.remove(report)} # start fresh every time you render
```

```{r}
data <- fread(file = here(experiment_path, "agriculture_yearly.csv"))

data <- data[cropnames != "Urban - residential, commercial, and industrial, unsegregated",] # remove urban
data$cropnames <- ifelse(data$cropnames %in% c("Unclassified fallow", "Idle"), "Fallow", data$cropnames) # rename unclassified
data <- filter(data, cropnames != "Fallow")

# create clusters based on location. Function creates clusters of size dist km. 
data$cluster <- mapply(assign_cluster, data$x, data$y, 75)

# control for variation across years
data <- year_control(data)
```

I then need to get the total water ETd
Since it's in mm/month, I need to multiply by the number of months
```{r}
data$ag_ET_year <- data$ag_ET*12
```

Use a clustered standard errors regression to retrieve the water consumption by county. 
This is specifically for the average over all the years (or my year of interest) so I can get county specific confidence intervals. 
```{r}
# average over years/select your year of choice
if (!is.na(ET_year)){
  data_sub <- filter(data, year == ET_year)
} else {
  data_sub <- data %>% group_by(x, y, NAME, Soil, cluster) %>% summarize(
    ag_ET_year = mean(ag_ET_year),
    ag_ET = mean(ag_ET),
    PET = mean(PET),
    ET_pred = mean(ET_pred),
    ET = mean(ET))
}

# get confidence intervals using clustered standard errors
ag_ET <- felm(ag_ET_year~NAME-1 | 0 | 0 | cluster, data_sub) 

# extract the mean and ci values from each
df_from_felm <- function(input_lm){
  ET_df <- as.data.frame(cbind(input_lm$coefficients, input_lm$cse))
  colnames(ET_df) <- c("ag_ET_mm_year", "ag_ET_se")
  ET_df$ag_ET_min <- ET_df$ag_ET_mm_year-(2*ET_df$ag_ET_se)
  ET_df$ag_ET_max <- ET_df$ag_ET_mm_year+(2*ET_df$ag_ET_se)
  ET_df$NAME <- substring(rownames(ET_df), 5)
  return(ET_df)
}

ET_df <- df_from_felm(ag_ET)
```


Add back in some more information about the counties
```{r}
# average over county
county_avg <- data_sub %>%
  group_by(NAME) %>% # group by county here
  summarize(
    lon = mean(x, na.rm=TRUE),
    lat = mean(y, na.rm=TRUE),
    soil = mean(Soil, na.rm=TRUE),
    ET = mean(ET, na.rm=TRUE), # all these are mm/day
    ET_pred = mean(ET_pred, na.rm=TRUE),
    ag_ET = mean(ag_ET, na.rm=TRUE),
    PET = mean(PET, na.rm=TRUE),
    n = n()) #get the number of pixels in each county

# add other information about the counties back in 
ET_df <- merge(ET_df, county_avg, by = "NAME")
```

Get a separate dataset of the ag ET by county over different years, but no confidence intervals this time. 
```{r}
all_years <- data %>%
  group_by(NAME, year) %>% # group by county here
  summarize(
    ag_ET_mm_year = mean(ag_ET_year, na.rm = TRUE),
    lon = mean(x, na.rm=TRUE),
    lat = mean(y, na.rm=TRUE),
    soil = mean(Soil, na.rm=TRUE),
    ET = mean(ET, na.rm=TRUE), # all these are mm/day
    ET_pred = mean(ET_pred, na.rm=TRUE),
    ag_ET = mean(ag_ET, na.rm=TRUE),
    PET = mean(PET, na.rm=TRUE),
    n = n())
```

## USGS Irrigation data

### Read in USGS data. Make a single irrigation dataset with 2010 and 2015 data
```{r}
irrigation2015 <- read.csv(here(usgs_irr_path,
                       "2015",
                       "water_use"), sep = "\t")[-1,] #first row just tells you the size of the entries; remove

irrigation2010 <- read.csv(here(usgs_irr_path,
                       "2010",
                       "water_use"), sep = "\t")[-1,] #first row just tells you the size of the entries; remove

irrigation2015$irryear <- 2015
irrigation2010$irryear <- 2010

irrigation <- rbind(irrigation2015, irrigation2010)

rm(irrigation2010)
rm(irrigation2015)
```

### Clean USGS irrigation data
Change units to kg/day and meters squared
```{r}
irrigation$NAME <- str_remove(irrigation$county_nm, " County") # remove the "County" suffix to each county name in order to match to county shapefile names


# convert billions of gallons per day to kg per day
conversion <- function(Mgal.d){
  Mgal.d = as.numeric(Mgal.d)
  kg.d = Mgal.d * 3.785411784 * 1000000
}

to_convert = names(irrigation)[c(6:10, 15)] #the names of the columns that are in Mgal/d and need to be converted
irrigation <- mutate(irrigation, across(to_convert, conversion))


# convert thousands of acres to square meters
conversion <- function(tacre){
  tacre = as.numeric(tacre)
  m2 = tacre * 1000 * 4046.86 # conversion to acres and conversion to square meters
}

to_convert = names(irrigation)[c(11:14)] #the names of the columns that are in thousand of acres and need to be converted

irrigation <- mutate(irrigation, across(to_convert, conversion))

# get a kg/year variable to make the plot more intuitive
irrigation$kg.year = irrigation$Irrigation..Crop.total.self.supplied.withdrawals.for.crops..fresh..in.Mgal.d*365

# add a mm/day and mm/year variable
# keep in mind the column names no longer reflect the true units
irrigation$mm.day <- irrigation$Irrigation..Crop.total.self.supplied.withdrawals.for.crops..fresh..in.Mgal.d/irrigation$Irrigation..Crop.total.irrigation.for.crops..in.thousand.acres

irrigation$mm.year <- irrigation$mm.day*365

# also get the percent drip and flood irrigation for each county
irrigation$drip <- irrigation$Irrigation..Crop.microirrigation.for.crops..in.thousand.acres/irrigation$Irrigation..Crop.total.irrigation.for.crops..in.thousand.acres
irrigation$flood <- irrigation$Irrigation..Crop.surface.irrigation.for.crops..in.thousand.acres/irrigation$Irrigation..Crop.total.irrigation.for.crops..in.thousand.acres
irrigation$sprinkler <- irrigation$Irrigation..Crop.sprinkler.irrigation.for.crops..in.thousand.acres/irrigation$Irrigation..Crop.total.irrigation.for.crops..in.thousand.acres
```


```{r}

if (is.na(irr_year)){
  # take the average irrigation over different years to simplify
  irrigation_avg <- irrigation %>% 
    group_by(NAME) %>% 
    summarize(mm.year = mean(mm.year),
              kg.year = mean(kg.year),
              mm.day = mean(mm.day),
              flood = mean(flood), 
              drip = mean(drip), 
              sprinkler = mean(sprinkler))
} else {
  # or just use one year
  irrigation_avg <- irrigation %>% filter(irryear == irr_year)
  irrigation_avg$year <- NULL
}

```

```{r}
# merge ET and irrigation datasets to get the irrigation efficiency
efficiency_df <- base::merge(irrigation_avg, ET_df, by="NAME", all.x = TRUE)
efficiency_df$irrigation_efficiency <- efficiency_df$ag_ET_mm_year*100/efficiency_df$mm.year
efficiency_df$irrigation_efficiency_min <- efficiency_df$ag_ET_min*100/efficiency_df$mm.year
efficiency_df$irrigation_efficiency_max <- efficiency_df$ag_ET_max*100/efficiency_df$mm.year
```

## Do some plotting!

### Read in county shapefile and merge dfs
```{r}
# study area and counties
study_area <- st_read(study_area_loc) %>% st_transform(st_crs("+proj=longlat +datum=WGS84"))
counties <- st_read(counties_loc) %>% filter(STATEFP == "06") %>% st_transform(st_crs("+proj=longlat +datum=WGS84"))
counties_irrigation <- base::merge(counties, efficiency_df, by = "NAME")
```


### Plot 1: irrigation use and agricultural water consumption
```{r, eval = TRUE}
plotted_datasets <- ggplot() + 
  geom_sf(data = counties_irrigation, aes(fill = kg.year/1000000000), color=alpha("black",.2), size = .2) + # total water diversions in Tg
  scale_fill_gradient2(name="Irrigation (kq/year)", low = "red", mid = "white", high = "deepskyblue4", breaks = c(1000, 2000)) + 
  # geom_sf(data = counties_irrigation, fill=alpha("red",0), color=alpha("grey64")) + 
  geom_sf(data = study_area, fill=alpha("red",0), color = "black", size = .2) + 
  new_scale_fill() +
  geom_raster(data = data[sample(nrow(data), size = nrow(data), replace = FALSE),], aes(x=x, y=y, fill=ag_ET_year), alpha=1) +
  scale_fill_gradientn(name="Agricultural ET (mm/year)", colours = c("white", "#FFFFCC", "#A1DAB4",  "#41B6C4", "#2C7FB8", "#253494", "#081D58", "#081D58"), limits = c(-391, 2346), breaks = seq(0, 2000, by = 1000)) +
  theme_classic() +
  scale_x_continuous(name = "Longitude", 
                     limits = c(minlong, maxlong),
                     breaks = round(seq(minlong + 1, maxlong + 1, by = 3))) + 
  scale_y_continuous(name = "Latitude", 
                     limits = c(minlat, maxlat),
                     breaks = round(seq(minlat + 1, maxlat + 1, by = 2))) + 
  annotation_scale() + 
  theme(legend.position = c(.8, .9), 
        legend.direction = "horizontal", 
        axis.text.x = element_text(size = 6),
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7),
        legend.title = element_blank(),
        legend.margin = margin(18))

plotted_datasets
```

### Plot 2: irrigation use and agricultural water consumption in the central valley
```{r}
# remove counties that don't have any CV agriculture
cv_data <- counties_irrigation
cv_data <- filter(counties_irrigation, n > 50000)

plotted_eff <- ggplot() + 
  geom_sf(data = counties, fill = "grey", color=alpha("white",1), size = .2) + 
  geom_sf(data = cv_data, aes(fill = irrigation_efficiency), color=alpha("black",.2), size = .2) + 
  geom_text(data = cv_data, aes(x = lon, y = lat, label = round(irrigation_efficiency, 0)), size = 2.2) + 
  scale_fill_gradient2(name = "Irrigation efficiency (%)", low = "red", mid = "white", high = "#70AD47", limits = c(0, 100)) + # alternatve top limit: max(cv_data$irrigation_efficiency)
  geom_sf(data = study_area, fill=alpha("red",0), color = "black", size = .2) + 
  xlim(c(minlong, maxlong)) + 
  ylim(c(minlat, maxlat)) + 
  theme_classic() +
  scale_x_continuous(name = "Longitude", 
                     limits = c(minlong, maxlong),
                     breaks = round(seq(minlong + 1, maxlong + 1, by = 3))) + 
  scale_y_continuous(name = "", 
                     limits = c(minlat, maxlat),
                     breaks = c()) + 
  theme(legend.position = c(.805, .85), 
        legend.direction = "horizontal", 
        axis.text.x = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        legend.title = element_text(size = 7),# Set legend title size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7),
        axis.text.y = element_blank(),    # Remove y-axis text
        axis.ticks.y = element_blank(),   # Remove y-axis ticks
        axis.line.y = element_blank()) + 
  guides(fill = guide_colourbar(title.position="top", title.hjust = 0.5))#, override.aes = list(size = 0.5))) 

plotted_eff
```

```{r}
theme_set(theme_pubr())

figure <- ggarrange(plotted_datasets, plotted_eff, 
                    # labels = c("a", "b"),
                    ncol = 2, nrow = 1)
figure

ggsave(here(experiment_path, "Figure_4.pdf"), 
       width = 18,
       height = 12.2,
       units = "cm")
```


# Look at uncertainty based on what you get with combining different years
```{r}
# mix and match all years to get irrigation efficiencies
combo <- data.table()
for (ag_y in c(2016, 2018, 2019)){
  for (irr_y in c(2010, 2015)){
    irr <- irrigation %>% filter(irryear == irr_y)
    ag <- all_years %>% filter(year == ag_y)
    combined <- base::merge(irr, ag, by="NAME", all.x = TRUE)
    combined$irrigation_efficiency <- combined$ag_ET_mm_year*100/combined$mm.year
    combo <- rbind(combo, combined)
  }
}
combo <- filter(combo, n > 50000)
```


```{r}
cv_data$none <- ""
ggplot(combo, aes(x = reorder(NAME, -lat))) + 
  geom_point(data = combo, aes(y = irrigation_efficiency, shape = as.factor(year.y), color = as.factor(irryear)), alpha = .7, size = 2) + 
  geom_point(data = cv_data, aes(y = irrigation_efficiency, fill = none), color = "red", size = 3, shape = "cross") + 
  # geom_linerange(data = cv_data, aes(ymin = irrigation_efficiency_min, ymax = irrigation_efficiency_max, fill = none), color = "red") + 
  theme_bw() + 
  theme(legend.position = c(.25, .75), 
        axis.text.x = element_text(angle = 30, hjust=1, size = 7), 
        legend.direction = "horizontal", 
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        legend.title = element_text(size = 7),# Set legend title size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7)) + 
  xlab("County") +
  ylab("Irrigation efficiency (%)") + 
  guides(shape=guide_legend(title="Agricultural ET year"), 
         color=guide_legend(title="Irrigation year"), 
         fill=guide_legend(title="Average over all years"))

ggsave(here(experiment_path, "Figure_S8.pdf"), 
       width = 16,
       height = 10,
       units = "cm")
```

# Compare predictions to what the data returned
```{r}
efficiency_calc <- function(flood, drip){
  sprinkler = 1-flood-drip
  
  # conveyance efficiency
  CE = .85 # theoretically depends on type of conveyance technology
  
  # distribution/management efficiency
  DE = .95 
  
  # application efficiency
  AE = (drip*.95) + (sprinkler*.75) + (flood*.60)
  
  return(CE*DE*AE*100)
}

counties_irrigation$theoretical_efficiency <- efficiency_calc(counties_irrigation$flood,  counties_irrigation$drip)
```

```{r}
# plot
counties_irrigation <-  filter(counties_irrigation, n > 50000)
counties_irrigation %>%
  ggplot() + 
  geom_smooth(aes(x = theoretical_efficiency, y = irrigation_efficiency), method = lm) + 
   geom_linerange(aes(x = theoretical_efficiency, ymin = irrigation_efficiency_min, ymax = irrigation_efficiency_max)) + 
  geom_abline(intercept=0,slope=1, color="red") + 
  geom_point(aes(x = theoretical_efficiency, y = irrigation_efficiency, color=lat), size = 1.5) +
  scale_color_distiller(palette="RdYlBu", direction=1, name = "Latitude") + 
  theme_bw() + 
  coord_fixed() + 
  xlab("Theoretical irrigation efficiency (%)") + 
  ylab("Empirical irrigation efficiency (%)") +
  xlim(c(min(counties_irrigation$theoretical_efficiency, counties_irrigation$irrigation_efficiency), max(counties_irrigation$theoretical_efficiency, counties_irrigation$irrigation_efficiency))) +
  ylim(c(min(counties_irrigation$theoretical_efficiency, counties_irrigation$irrigation_efficiency), max(counties_irrigation$theoretical_efficiency, counties_irrigation$irrigation_efficiency))) +
  theme(axis.text.x = element_text(size = 6),
        axis.text.y = element_text(size = 6),
        strip.text = element_text(size = 7),
        legend.text = element_text(size = 6), # Set legend text size to 7pt
        legend.title = element_text(size = 7),# Set legend title size to 7pt
        axis.title.x = element_text(size = 7), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 7))

ggsave(here(experiment_path, "Figure_S7.pdf"), 
       width = 12,
       height = 11,
       units = "cm")
```

## Statsitics: overall efficiency + CIs
```{r}
# In order to get confidence intervals for the overall numbers I will need the entire dataset
data_irr <- merge(data, irrigation_avg, by = "NAME", all.x = TRUE, all.y = FALSE)
data_irr$irrigation_efficiency <- data_irr$ag_ET_year*100/data_irr$mm.year

# get confidence intervals using clustered standard errors
total_eff <- felm(irrigation_efficiency~1 | 0 | 0 | cluster, data_irr) #+year

# extract the mean and ci values from each
df_from_felm <- function(input_lm){
  ET_df <- as.data.frame(cbind(input_lm$coefficients, input_lm$cse))
  colnames(ET_df) <- c("mean", "se")
  ET_df$min <- ET_df$mean-(2*ET_df$se)
  ET_df$max <- ET_df$mean+(2*ET_df$se)
  ET_df$NAME <- substring(rownames(ET_df), 5)
  return(ET_df)
}

total_eff_df <- df_from_felm(total_eff)

write("Average irrigation efficiency",file=report,append=TRUE)
fwrite(total_eff_df,file=report,append=TRUE)
```
