==> code/3_analysis/6_CalSIMETAW_comparison.Rmd <==
---
title: "CalSIMETAW comparison"
author: "Anna Boser"
date: "2022-12-20"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(data.table)
library(here)
library(dplyr)
library(ggplot2)
library(tidyr)
library(stringr)
library(ggrepel)
library(lfe)

source(here("file_paths.R"))
source(here("helper_functions.R"))
```

```{r}
experiment_name <- "fallow0.05,2_4-18_gb"
experiment_path <- here("data", "4_for_analysis", "ML_outputs",  "experiments", experiment_name)

report <- here(experiment_path, "6_report.txt")
if (file.exists(report)) {file.remove(report)} # start fresh every time you render
```


# Diagnostics
```{r}
# disgnostics: is simulated natural ET higher any given month? 
c <- fread(here(experiment_path, "time_series_by_crop.csv"))
sc <- fread(here(experiment_path, "time_series_by_subcrop.csv"))

ggplot(c) + 
  geom_line(aes(x = month, y = ET, color = "Total ET")) + 
  geom_line(aes(x = month, y = ET_pred, color = "Simulated natural ET")) + 
  facet_wrap(vars(cropnames))

ggplot(sc) + 
  geom_line(aes(x = month, y = ET, color = "Total ET")) + 
  geom_line(aes(x = month, y = ET_pred, color = "Simulated natural ET")) + 
  facet_wrap(vars(subcropnames))
```

## Compare my estimates of crop water use and irrigation water use to CalSIMETAW outputs
Data from: https://data.ca.gov/dataset/cal-simetaw-unit-values

```{r}
CS <- fread(calsimetaw_loc)

# keep only counties of interest
counties <- c("Tehama", "Glenn", "Butte", "Colusa", "Sutter", "Yuba", "Yolo", "Solano", "Sacramento", "San Joaquin", "Stanislaus", "Merced", "Madera", "Fresno", "Kings", "Tulare", "Kern")

CS <- filter(CS, COUNTY_NAME %in% counties)

# original values are in mm/month: https://data.ca.gov/dataset/cal-simetaw-unit-values/resource/10b6414f-c281-4a23-8204-711f867b5109
CS$Date <- as.Date(CS$Date)
CS$year <- year(CS$Date + 92) # add three months to get the water year
CS$month <- month(CS$Date)

# subcrop lookup table. The rule is first look in the general crop types, if it's not there look in the subcrop types. 
l = c("Grain" = "Grain and hay crops", 
      "Rice" = "Rice", 
      "Cotton" = "Cotton", 
      "Sugar Beets" = NA, # "Truck, nursery, and berry crops", 
      "Corn" = "Corn,Sorghum or Sudan", # "Field crops",
      "Dry Beans" = "Beans (dry)", # "Truck, nursery, and berry crops",
      "Safflower" = "Safflower", # "Field crops",
      "Other Field" = "Field crops",
      "Alfalfa" = "Alfalfa & alfalfa mixtures", # "Grain and hay crops", 
      "Pasture" = "Pasture", 
      "Tomato Processing" = "Tomatoes (processing)", 
      "Tomato Fresh" = NA, # "Truck, nursery, and berry crops", 
      "Cucurbits" = "Melons", # "Truck, nursery, and berry crops", 
      "Onions & Garlic" = "Onions & garlic", # "Truck, nursery, and berry crops", 
      "Potatoes" = "Potato or Sweet potato", # "Truck, nursery, and berry crops", 
      "Truck Crops" = "Truck, nursery, and berry crops", 
      "Almond & Pistacios" = "Almond & Pistacios", # "Deciduous fruits and nuts", 
      "Other Decidious" = "Deciduous fruits and nuts", 
      "Citrus & Subtropical" = "Citrus and subtropical", 
      "Vineyard" = "Vineyards", 
      "Urban Landscape" = NA, 
      "Riparian" = NA, 
      "Native Vegetation" = NA, 
      "Water Surface" = NA)

CS$cropnames = l[CS$CROP_NAME]
CS <- filter(CS, !is.na(cropnames))
```

```{r}
# get the monthly ET by crop dataset
df <- fread(here(experiment_path, "agriculture_monthly_county_crop.csv"))
df <- filter(df, cropnames %in% CS$cropnames)
CS <- merge(CS, df, by = c("cropnames", "COUNTY_NAME", "month"))

scalar = 12 # mm/month to mm/year
# average by CalSIMETAW county and crop type.
scatter_data <- CS %>% 
  group_by(COUNTY_NAME, cropnames) %>% 
  summarize(ETo = mean(ETo)*scalar, 
            ag_ET_need = mean(ifelse(ETc>Pcp, ETc-Pcp, 0))*scalar, 
            ETc_ET_pred = mean(ifelse(ETc>ET_pred, ETc-ET_pred, 0))*scalar, #
            ET_ET_pred = mean(ifelse(ET>ET_pred, ET-ET_pred, 0))*scalar, #
            ag_ET = mean(ag_ET)*scalar, #
            ET = mean(ET)*scalar, #
            ET_pred = mean(ET_pred)*scalar, #
            latitude = mean(latitude), #
            unused_rain = mean(ifelse(ETc<Pcp, Pcp-ETc, 0))*scalar,
            ETc = mean(ETc)*scalar, 
            Pcp = mean(Pcp)*scalar, 
            Er = mean(Er)*scalar, 
            Spg = mean(Spg)*scalar, 
            Espg = mean(Espg)*scalar, 
            AW = mean(AW)*scalar)
```

# simetaw-ET_pred (monthly) vs ag ET (monthly)
```{r}
lm(ETc_ET_pred~ET_ET_pred, scatter_data) %>% summary()
line <- paste("correlation for simetaw-ET_pred (monthly) vs ag ET (monthly):", cor(scatter_data$ETc_ET_pred,scatter_data$ET_ET_pred))
write(line,file=report,append=TRUE)

bias <- mean(scatter_data$ETc_ET_pred - scatter_data$ET_ET_pred)
se <- sd(scatter_data$ETc_ET_pred - scatter_data$ET_ET_pred)/sqrt(nrow(scatter_data))
line <- print_CI("bias simetaw-ET_pred (monthly) vs ag ET (monthly)", bias, se)
print(line)
write(line,file=report,append=TRUE)

line <- print_CI("bias as percent of theoretical", (bias/mean(scatter_data$ETc_ET_pred))*100, (se/mean(scatter_data$ETc_ET_pred))*100)
print(line)
write(line,file=report,append=TRUE)
```


# simetaw-rain (monthly) vs ag ET (monthly)
```{r}
lm(ag_ET_need~ET_ET_pred, scatter_data) %>% summary()
line <- paste("correlation for simetaw-rain (monthly) vs ag ET (monthly):", cor(scatter_data$ag_ET_need,scatter_data$ET_ET_pred))
write(line,file=report,append=TRUE)

bias <- mean(scatter_data$ag_ET_need - scatter_data$ET_ET_pred)
se <- sd(scatter_data$ag_ET_need - scatter_data$ET_ET_pred)/sqrt(nrow(scatter_data))
line <- print_CI("bias simetaw-rain (monthly) vs ag ET (monthly)", bias, se)
print(line)
write(line,file=report,append=TRUE)

line <- print_CI("bias as percent of theoretical", (bias/mean(scatter_data$ag_ET_need))*100, (se/mean(scatter_data$ag_ET_need))*100)
print(line)
write(line,file=report,append=TRUE)
```


# combine the two plots I want into one
```{r}
cols <- c("#89C5DA", "#DA5724", "#74D944", "#CE50CA", "#3F4921", "#C0717C", "#CBD588", "#5F7FC7", 
"#673770", "#D3D93E", "#38333E", "#508578", "#D7C1B1", "#689030", "#AD6F3B", "#CD9BCD", 
"#D14285", "#6DDE88")
fin_data <- pivot_longer(scatter_data, cols = c("ag_ET_need", "ETc_ET_pred"), names_to = "method", values_to = "CalSIMETAW")
fin_data$method <- factor(fin_data$method, levels = c("ETc_ET_pred", "ag_ET_need"))

new.labs <- c("Naturally-occurring ET as counterfactual", "Precipitation as counterfactual")
names(new.labs) <- c("ETc_ET_pred", "ag_ET_need")

ggplot(fin_data) + 
  geom_abline(intercept=0,slope=1, color="red") + 
  geom_point(aes(x = ET_ET_pred, y = CalSIMETAW, color = cropnames)) + 
  scale_color_manual(values=cols) + 
  geom_smooth(aes(x = ET_ET_pred, y = CalSIMETAW), method="lm") + 
  facet_grid(cols = vars(method), labeller = labeller(method = new.labs)) + 
  theme_classic() + 
  ylab("CalSIMETAW Agricultural ET (mm/year)") + 
  xlab("Empirical Agricultural ET (mm/year)") + 
  coord_fixed() + 
  xlim(c(min(fin_data$CalSIMETAW, fin_data$ET_ET_pred), max(fin_data$CalSIMETAW, fin_data$ET_ET_pred))) +
  ylim(c(min(fin_data$CalSIMETAW, fin_data$ET_ET_pred), max(fin_data$CalSIMETAW, fin_data$ET_ET_pred))) +
  guides(color=guide_legend(title="Crop type")) + 
  theme(axis.text.x = element_text(size = 10.5),
        axis.text.y = element_text(size = 10.5),
        strip.text = element_text(size = 10.5),
        legend.text = element_text(size = 10.5), # Set legend text size to 7pt
        legend.title = element_text(size = 10.5),# Set legend title size to 7pt
        axis.title.x = element_text(size = 10.5), # Set x-axis title size to 7pt
        axis.title.y = element_text(size = 10.5)) # because I have to save as size 27 instead of 18, 7*1.5 = 10.5 which will appear 7

ggsave(here(experiment_path, "Figure_S6.pdf"), 
       width = 27,
       height = 12,
       units = "cm")
```
